requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
astroid==3.3.11
asttokens==3.0.0
async-lru==2.0.5
asyncssh==2.21.0
atomicwrites==1.4.1
attrs==25.3.0
autopep8==2.0.4
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==45.0.6
curl_cffi==0.13.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
dill==0.4.0
discord.py==2.6.4
docstring-to-markdown==0.17
docutils==0.21.2
executing==2.2.0
fastjsonschema==2.21.2
ffmpeg==1.4
flake8==7.1.2
fonttools==4.59.2
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jellyfish==1.2.0
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==25.6.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
multitasking==0.0.12
murmurhash==1.0.13
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
nose==1.3.7
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
numpydoc==1.9.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
peewee==3.18.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.33.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyamg==5.3.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydocstyle==6.3.0
pyflakes==3.2.0
pygame==2.6.1
PyGithub==2.7.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.8
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.2
pyuca==1.2
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.4.0
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
roman-numerals-py==3.1.0
rope==1.14.0
rpds-py==0.27.1
rtree==1.4.1
scikit-learn==1.7.1
scipy==1.16.1
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.0.8
spyder-kernels==3.0.5
srsly==2.5.1
stack-data==0.6.3
superqt==0.7.6
sympy==1.14.0
terminado==0.18.1
text-unidecode==1.3
textdistance==4.6.3
thinc==8.3.6
threadpoolctl==3.6.0
three-merge==0.1.1
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.1
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uri-template==1.3.0
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
whatthepatch==1.0.7
wheel==0.45.1
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.3
yapf==0.43.0
yarl==1.20.1
yfinance==0.2.66
zipp==3.23.0
